Move util functions out of controllers/base
