+++ indico/modules/events/abstracts/controllers/reviewing.py
@@ -25,10 +25,9 @@ from indico.modules.events.abstracts.controllers.base import (RHAbstractBase, RH
                                                               AbstractsExportPDFMixin,
                                                               AbstractsDownloadAttachmentsMixin, AbstractsExportCSV,
                                                               AbstractsExportExcel,
-                                                              CustomizeAbstractListMixin, build_review_form,
-                                                              render_abstract_page)
+                                                              CustomizeAbstractListMixin)
 from indico.modules.events.abstracts.forms import (AbstractCommentForm, AbstractJudgmentForm,
-                                                   AbstractReviewedForTracksForm)
+                                                   AbstractReviewedForTracksForm, build_review_form)
 from indico.modules.events.abstracts.models.abstracts import Abstract, AbstractState
 from indico.modules.events.abstracts.models.comments import AbstractComment
 from indico.modules.events.abstracts.operations import (judge_abstract, reset_abstract_state, withdraw_abstract,
@@ -36,7 +35,7 @@ from indico.modules.events.abstracts.operations import (judge_abstract, reset_ab
                                                         update_abstract_comment, create_abstract_review,
                                                         update_abstract_review, update_reviewed_for_tracks)
 from indico.modules.events.abstracts.util import (AbstractListGeneratorDisplay, get_track_reviewer_abstract_counts,
-                                                  get_user_tracks)
+                                                  get_user_tracks, render_abstract_page)
 from indico.modules.events.abstracts.views import WPDisplayAbstractsReviewing
 from indico.modules.events.tracks.models.tracks import Track
 from indico.util.i18n import _
